requirements.txt
# Web 框架
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# 数据库
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pgvector>=0.3.0

# 认证
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
passlib[bcrypt]>=1.7.4

# Redis
redis[hiredis]>=5.0.0

# 配置
pydantic>=2.6.0
pydantic-settings>=2.2.0

# AI 模型
google-genai>=0.2.0

# 向量嵌入
sentence-transformers>=3.0.0

# 邮件
aiosmtplib>=3.0.0
httpx>=0.27.0

# 工具
python-multipart>=0.0.9
Pillow>=10.0.0
pyyaml>=6.0.0
cuid2>=2.0.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0

# 代码质量
black>=24.0.0
pylint>=3.0.0
mypy>=1.8.0
